+++ NEWS.md
@@ -1,3 +1,7 @@
+## templates 0.0.3 2026_03_21
+
+* Add generic visualisation template
+
 ## templates 0.0.2 2026_02_03
 
 * Add blog post template